change some int64_t into long long

+++ src/rpc_global.cc
@@ -115,9 +115,9 @@ bool SRPCGlobal::task_init(RPCClientParams& params, ParsedURI& uri,
 	return false;
 }
 
-int64_t SRPCGlobal::get_trace_id()
+long long SRPCGlobal::get_trace_id()
 {
-	int64_t trace_id = 0;
+	long long trace_id = 0;
 	this->snowflake.get_id(0, 0, &trace_id);
 	return trace_id;
 }
