+++ src/rpc_span.h
@@ -56,15 +56,15 @@ public:
 		this->timestamp_shift = this->group_shift + this->group_bits;
 	}
 
-	bool get_id(int64_t group_id, int64_t machine_id, int64_t *uid)
+	bool get_id(long long group_id, long long machine_id, long long *uid)
 	{
 		if (group_id > this->group_id_max || machine_id > this->machine_id_max)
 			return false;
 
 		struct timespec ts;
 		clock_gettime(CLOCK_MONOTONIC, &ts);
-		int64_t timestamp = ts.tv_sec * 1000 + ts.tv_nsec / 1000000;
-		int64_t seq_id;
+		long long timestamp = ts.tv_sec * 1000 + ts.tv_nsec / 1000000;
+		long long seq_id;
 
 		if (timestamp < this->last_timestamp)
 			return false;
@@ -89,21 +89,21 @@ public:
 	}
 
 private:
-	std::atomic<int64_t> last_timestamp; // -1L;
-	std::atomic<int64_t> sequence; // 0L;
+	std::atomic<long long> last_timestamp; // -1L;
+	std::atomic<long long> sequence; // 0L;
 
 	int timestamp_bits;
 	int group_bits;
 	int machine_bits;
 	int sequence_bits;
 
-	int64_t group_id_max;
-	int64_t machine_id_max;
-	int64_t sequence_max;
+	long long group_id_max;
+	long long machine_id_max;
+	long long sequence_max;
 
-	int64_t timestamp_shift;
-	int64_t group_shift;
-	int64_t machine_shift;
+	long long timestamp_shift;
+	long long group_shift;
+	long long machine_shift;
 
 	// u_id = [timestamp][group][machine][sequence]
 	static constexpr int TIMESTAMP_BITS = 37;
@@ -116,16 +116,16 @@ private:
 class RPCSpan
 {
 private:
-	int64_t trace_id;
-	int64_t span_id;
-	int64_t parent_span_id;
+	long long trace_id;
+	long long span_id;
+	long long parent_span_id;
 	std::string service_name;
 	std::string method_name;
 	int data_type;
 	int compress_type;
-	int64_t start_time;
-	int64_t end_time;
-	int64_t cost;
+	long long start_time;
+	long long end_time;
+	long long cost;
 	std::string remote_ip;
 	int state;
 	int error;
@@ -145,14 +145,14 @@ public:
 		remote_ip("")
 	{}
 
-	int64_t get_trace_id() const { return this->trace_id; }
-	void set_trace_id(int64_t id) { this->trace_id = id; }
+	long long get_trace_id() const { return this->trace_id; }
+	void set_trace_id(long long id) { this->trace_id = id; }
 
-	int64_t get_span_id() const { return this->span_id; }
-	void set_span_id(int64_t id) { this->span_id = id; }
+	long long get_span_id() const { return this->span_id; }
+	void set_span_id(long long id) { this->span_id = id; }
 
-	int64_t get_parent_span_id() const { return this->parent_span_id; }
-	void set_parent_span_id(int64_t id) { this->parent_span_id = id; }
+	long long get_parent_span_id() const { return this->parent_span_id; }
+	void set_parent_span_id(long long id) { this->parent_span_id = id; }
 
 	const std::string& get_service_name() const { return this->service_name; }
 	void set_service_name(std::string name) { this->service_name = name; }
@@ -166,14 +166,14 @@ public:
 	int get_compress_type() const { return this->compress_type; }
 	void set_compress_type(int type) { this->compress_type = type; }
 
-	int64_t get_start_time() const { return this->start_time; }
-	void set_start_time(int64_t time) { this->start_time = time; }
+	long long get_start_time() const { return this->start_time; }
+	void set_start_time(long long time) { this->start_time = time; }
 
-	int64_t get_end_time() const { return this->end_time; }
-	void set_end_time(int64_t time) { this->end_time = time; }
+	long long get_end_time() const { return this->end_time; }
+	void set_end_time(long long time) { this->end_time = time; }
 
-	int64_t get_cost() const { return this->cost; }
-	void set_cost(int64_t time) { this->cost = time; }
+	long long get_cost() const { return this->cost; }
+	void set_cost(long long time) { this->cost = time; }
 
 	const std::string& get_remote_ip() const { return this->remote_ip; }
 	void set_remote_ip(std::string ip) { this->remote_ip = std::move(ip); }
